refactor(driver): refactor score driver
